use slug as id for fetching collectibles
